service/tcfd-service/requirements.txt
# TCFD Service Requirements
# MSV Pattern with Layered Architecture

# FastAPI 및 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 데이터베이스 연결
asyncpg==0.29.0
psycopg2-binary==2.9.9

# 데이터 검증 및 직렬화
pydantic==2.5.0
pydantic-settings==2.1.0

# 환경 변수 관리
python-dotenv==1.0.0

# 로깅 및 유틸리티
python-json-logger==2.0.7

# AI 및 머신러닝 (기존 TCFD 기능)
langchain==0.0.350
openai==1.3.7
chromadb==0.4.18
sentence-transformers==2.2.2

# 벡터 데이터베이스
chromadb==0.4.18

# HTTP 클라이언트
httpx==0.25.2
aiohttp==3.9.1

# 비동기 처리
asyncio-mqtt==0.16.1

# 데이터 처리
pandas==2.1.4
numpy==1.25.2

# 테스트
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# 개발 도구
black==23.11.0
isort==5.12.0
flake8==6.1.0
